+++ src/lib/src/service/lib.service.ts
@@ -88,4 +88,32 @@ export class LocalStorage {
 
   }
 
+  /**
+   * Sets an item in local storage, and auto-subscribes
+   * @param key The item's key
+   * @param data The item's value, must NOT be null or undefined
+   */
+  setItemSubscribe(key: string, data: any): void {
+
+    this.setItem(key, data).subscribe(() => {}, () => {});
+
+  }
+
+  /**
+   * Deletes an item in local storage, and auto-subscribes
+   * @param key The item's key
+   */
+   removeItemSubscribe(key: string): void {
+
+    this.removeItem(key).subscribe(() => {}, () => {});
+
+  }
+
+  /** Deletes all items from local storage, and auto-subscribes */
+  clearSubscribe(): void {
+
+    this.clear().subscribe(() => {}, () => {});
+
+  }
+
 }